pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basic-chat"
version = "1.0.0"
description = "AI Chat Application with Advanced Reasoning Capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Souriya Khaosanga", email = "souriya@chainable.ai"}
]
maintainers = [
    {name = "Souriya Khaosanga", email = "souriya@chainable.ai"}
]
keywords = ["ai", "chat", "streamlit", "ollama", "reasoning", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.340",
    "langchain-community>=0.0.11",
    "langchain-experimental>=0.0.40",
    "langchain-core>=0.1.0",
    "langchain-ollama>=0.1.0",
    "pydantic>=2.0.0",
    "gTTS>=2.3.2",
    "pytz>=2023.3",
    "chromadb>=0.4.0",
    "pillow>=10.0.0",
    "pypdf>=3.0.0",
    "unstructured>=0.10.0",
    "sentence-transformers>=2.2.0",
    "duckduckgo-search>=4.1.1",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    "redis>=4.5.0",
    "cachetools>=5.3.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
basic-chat = "src.core.app:main"

[project.urls]
Homepage = "https://github.com/khaosans/basic-chat-template"
Documentation = "https://github.com/khaosans/basic-chat-template/tree/main/docs"
Repository = "https://github.com/khaosans/basic-chat-template"
Issues = "https://github.com/khaosans/basic-chat-template/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.sql"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
] 